python-services/ocr-service/requirements_ocr.txt
Flask==3.0.0
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.10.0.84
